Tag组件的逻辑转为useTags文件
[Feature]
Tag组件中的逻辑代码移动到useTag.js。
Tasks组件中的Today's tasks区域的task将显示tag。
